requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
openpyxl>=3.1.2
google-cloud-storage>=2.16.0
xlrd>=2.0.1
